requirements.txt
openai
python-dotenv
diffusers
transformers
compel
fastapi
accelerate
# CUDA 12.6 対応 PyTorch公式ホイール
--index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio